feat: modify vertify code structure
