+++ src/controllers/userController.ts
@@ -4,13 +4,13 @@ import AppException from '../exceptions/appException';
 import { HttpStatusCode } from 'axios';
 
 export const userController = {
-    deleteUser: async (req: Request, res: Response, next: NextFunction) => {
-        const { userId } = req.params;
-        if (!userId) {
-            return next(new AppException(HttpStatusCode.BadRequest, 'Missing userId parameter'));
-        }
+  deleteUser: async (req: Request, res: Response, next: NextFunction) => {
+    const { userId } = req.params;
+    if (!userId) {
+      return next(new AppException(HttpStatusCode.BadRequest, 'Missing userId parameter'));
+    }
 
-        await userService.deleteUserById(userId);
-        return res.status(200).json({ message: 'User and related memberships deleted' });
-    },
+    await userService.deleteUserById(userId);
+    return res.status(200).json({ message: 'User and related memberships deleted' });
+  },
 };
